Move OOM setup to test suite.
